requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
deprecation==2.1.0
Django==4.2.17
django-cors==0.1
django-cors-headers==4.6.0
django-eventstream==4.5.1
django-grip==3.5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
gripcontrol==4.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
MarkupSafe==3.0.2
packaging==24.2
postgrest==0.19.1
psycopg2-binary==2.9.10
pubcontrol==3.5.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
StrEnum==0.4.15
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
